added callback from Win32Window to Presenter (informing about window size changes); fixed bug with quiting while losing active status from window

+++ graphics/DxPresenter.hpp
@@ -43,10 +43,7 @@ public:
 	ptr<RenderBuffer> GetBackBuffer();
 	void SetMode(const PresentMode& mode);
 	void Present();
-
-	/// Изменить размер области вывода.
-	/** Вызывается окном, когда его размер изменяется. */
-	void Resize(size_t width, size_t height);
+	void Resize(int width, int height);
 };
 
 END_INANITY_GRAPHICS
